chore(suite): refactor AccountException exports and filenames
