+++ packages/suite/src/components/wallet/AccountException/AccountNotExists.tsx
@@ -5,11 +5,9 @@ import { AccountExceptionLayout } from 'src/components/wallet';
  * Handler for invalid router params
  * see: @wallet-actions/selectedAccountActions
  */
-const AccountNotExists = () => (
+export const AccountNotExists = () => (
     <AccountExceptionLayout
         title={<Translation id="TR_ACCOUNT_EXCEPTION_NOT_EXIST" />}
         image="CLOUDY"
     />
 );
-
-export default AccountNotExists;
